[UPDATE] Input information
